:bug: Fix: itemPrice 오류 해결 ItemDetail price 추가 ItemListViewResponse price 타입 int 변경

+++ src/main/java/com/example/team1/Prometheus/entity/ItemListViewResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 public class ItemListViewResponse {
     private final String name;
-    private final double price;
+    private final int price;
 
     public ItemListViewResponse(ItemDetail itemDetail){
         this.name = itemDetail.getName();
